+++ src/Brassica/SoundChange.hs
@@ -1,7 +1,7 @@
 module Brassica.SoundChange (module X) where
 
 import Brassica.SoundChange.Apply     as X
-import Brassica.SoundChange.Category  as X
+import Brassica.SoundChange.Expand  as X
 import Brassica.SoundChange.Parse     as X
 import Brassica.SoundChange.Tokenise  as X
 import Brassica.SoundChange.Types     as X